django_farm_to_market/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg[binary]==3.1.18
pytest==7.4.3
black==23.11.0
flake8==6.1.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-filter==23.5 